+++ addon/macros/promise.js
@@ -23,7 +23,6 @@ export default function EmberCPM_promise(dependentKey) {
   return computed(dependentKey, function(){
     var value = get(this, dependentKey);
     if (value == null) { return value; }
-    // TODO: Use RSVP?
-    return Ember.$.when(value);
+    return Ember.RSVP.Promise.resolve(value);
   });
 }